pyproject.toml
[project]
name = "oldies-station"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.4",
    "loguru>=0.7.3",
    "python-decouple>=3.8",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.3",
]
